requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
duckdb==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-httpx==0.36.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
requests-mock==1.12.1
rich==14.3.3
six==1.17.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
